+++ packages/gamut-labs/src/landingPage/index.ts
@@ -5,3 +5,4 @@ export * from './PageHero';
 export * from './PageFeatures';
 export * from './Feature';
 export * from './PagePrefooter';
+export * from './PageVideoGallery';