+++ packages/gamut/src/Markdown/index.tsx
@@ -59,7 +59,7 @@ export class Markdown extends PureComponent<MarkdownProps> {
       overrides: userOverrides = {},
       skipDefaultOverrides = {},
       inline = false,
-      onAnchorClick = () => null,
+      onAnchorClick,
     } = this.props;
 
     if (!text) return null;
@@ -83,10 +83,19 @@ export class Markdown extends PureComponent<MarkdownProps> {
         }),
       !skipDefaultOverrides.a &&
         createTagOverride('a', {
-          component: (props: MarkdownAnchorProps) => (
-            <MarkdownAnchor onClick={onAnchorClick} {...props} />
-          ),
-          allowedAttributes: ['onClick'],
+          component: MarkdownAnchor,
+          processNode: (node, props) => {
+            // Note: this processNode override is necessary because wrapping this component
+            // in an anonymous functional component as with the Table below causes react rendering
+            // to crash with some chrome translation features.
+            // See https://codecademy.atlassian.net/browse/WEB-1214
+            return (
+              <MarkdownAnchor
+                onClick={onAnchorClick}
+                {...(props as MarkdownAnchorProps)}
+              />
+            );
+          },
         }),
       !skipDefaultOverrides.table &&
         createTagOverride('table', {