+++ packages/gamut/src/Toggle/index.tsx
@@ -1,7 +1,17 @@
-import cx from 'classnames';
-import React, { Component } from 'react';
+import { screenReaderOnly } from '@codecademy/gamut-styles';
+import styled from '@emotion/styled';
+import React from 'react';
 
-import styles from './styles/index.module.scss';
+import { Box } from '../Box';
+import { HiddenText } from '../HiddenText';
+
+export type ToggleSizes = keyof typeof sizes;
+export type ToggleVariants = typeof colors[number];
+
+export type LabelProps = {
+  disabled?: boolean;
+  variant?: ToggleVariants;
+};
 
 export type ToggleProps = {
   checked?: boolean;
@@ -9,48 +19,112 @@ export type ToggleProps = {
   onChange?: (...args: any[]) => any;
   label?: string;
   disabled?: boolean;
-  theme?: 'gray-blue' | 'purple';
-  size?: 'small' | 'medium';
+  variant?: ToggleVariants;
+  size?: ToggleSizes;
 };
 
-export class Toggle extends Component<ToggleProps, {}> {
-  render() {
-    const {
-      checked,
-      className,
-      onChange,
-      label,
-      disabled,
-      theme = 'gray-blue',
-      size = 'medium',
-    } = this.props;
-    return (
-      <label
-        className={cx(
-          styles.toggleButton,
-          {
-            [styles.toggled]: checked,
-            [styles.disabled]: disabled,
-          },
-          className
-        )}
-        arial-label={label}
-        htmlFor={label}
-      >
-        <input
-          type="checkbox"
-          checked={checked}
-          className={styles.invisible}
-          id={label}
-          disabled={disabled}
-          onChange={onChange}
-        />
-        <span className={styles.invisible}>{label}</span>
-        <div
-          className={cx(styles.track, styles[theme], styles[`track-${size}`])}
-        />
-        <div className={cx(styles.thumb, styles[`thumb-${size}`])} />
-      </label>
-    );
+const sizes = {
+  medium: {
+    height: '30px',
+    width: '60px',
+  },
+  small: {
+    height: '18px',
+    width: '36px',
+  },
+};
+
+const colors = ['blue', 'hyper'] as const;
+
+const ToggleTrack = styled(Box)`
+  transition: background-color 0.2s ease;
+
+  &:after {
+    content: '';
+    transition: opacity 0.2s ease;
+    opacity: 0;
+    border-radius: inherit;
+    position: absolute;
+    width: calc(100% + 8px);
+    height: calc(100% + 8px);
+    top: -4px;
+    left: -4px;
+    border-color: inherit;
+    border-style: solid;
+    border-width: 2px;
   }
-}
+
+  ${Box} {
+    transition: all 0.5s cubic-bezier(0.23, 1, 0.32, 1);
+  }
+`;
+
+const ToggleInput = styled.input(screenReaderOnly);
+
+const ToggleLabel = styled.label<LabelProps>`
+  display: inline-block;
+  cursor: pointer;
+  border: 0;
+  padding: 0;
+
+  &[disabled] {
+    cursor: not-allowed;
+    opacity: 0.75;
+  }
+
+  ${ToggleInput}:focus-visible + ${ToggleTrack} {
+    &:after {
+      opacity: 1;
+    }
+  }
+`;
+
+export const Toggle: React.FC<ToggleProps> = ({
+  checked,
+  className,
+  onChange,
+  label,
+  disabled,
+  variant = 'blue',
+  size = 'medium',
+}) => {
+  const activeColor = variant;
+  const checkedColor = checked ? variant : 'gray-500';
+  const sizeStyles = sizes[size];
+
+  return (
+    <ToggleLabel
+      className={className}
+      aria-label={label}
+      htmlFor={label}
+      variant={variant}
+      disabled={disabled}
+    >
+      <HiddenText>{label}</HiddenText>
+      <ToggleInput
+        type="checkbox"
+        checked={checked}
+        id={label}
+        disabled={disabled}
+        onChange={onChange}
+      />
+      <ToggleTrack
+        {...sizeStyles}
+        borderColor={activeColor}
+        backgroundColor={checkedColor}
+        borderRadius="99rem"
+        position="relative"
+      >
+        <Box
+          width="40%"
+          borderRadius="50%"
+          backgroundColor="white"
+          position="absolute"
+          top="10%"
+          bottom="10%"
+          left={checked ? '55%' : '5%'}
+        />
+      </ToggleTrack>
+    </ToggleLabel>
+  );
+};