+++ packages/styleguide/.storybook/components/Page/index.tsx
@@ -1,22 +1,24 @@
 import React, { useContext } from 'react';
-import { DocsContext, Title } from '@storybook/addon-docs/blocks';
+import { Description, DocsContext, Title } from '@storybook/addon-docs/blocks';
 import { Badge } from '../Badge';
 import { spacing } from '../styles';
 import { styled } from '@storybook/theming';
-import { theme } from '@codecademy/gamut-styles';
 import { OpenIcon } from '@codecademy/gamut-icons';
+import { Parameters } from '@storybook/addons';
+import { useKind } from '../TableOfContents/utils';
+import { Box, SectionLink } from '../TableOfContents/elements';
 
 const Link = styled.a`
   display: inline-flex;
   line-height: 1;
   column-gap: 0.5rem;
   align-items: flex-start;
-  color: ${theme.colors.blue};
+  color: ${({ theme }) => theme.color.secondary};
   border-bottom: 2px solid transparent;
 
   &:hover {
     text-decoration: none;
-    border-bottom-color: ${theme.colors.blue};
+    border-bottom-color: ${({ theme }) => theme.color.secondary};
   }
 `;
 
@@ -35,27 +37,67 @@ const HeaderRow = styled.div`
 
 const HeaderCol = styled.div``;
 
-interface Parameters {
-  component?: React.Component | string;
-  subcomponents?: Record<string, React.Component>;
-  pageTitle?: string;
+export interface GamutParameters extends Parameters {
+  subtitle?: string;
   status?: 'stable' | 'volatile' | 'deprecated';
   figmaId?: string;
   source?: string;
 }
 
-export const Page: React.FC = ({ children }) => {
-  const { parameters } = useContext(DocsContext);
-  const {
-    component,
-    subcomponents,
-    status,
-    pageTitle,
-    figmaId,
-    source,
-  }: Parameters = parameters;
+const BreadCrumbs: React.FC<{ path: string[] }> = ({ path }) => {
+  const { storyStore } = useContext(DocsContext);
+  if (path[0] === 'Gamut') return null;
 
-  const showStatus = Boolean(status || component || subcomponents);
+  const links = [{ text: 'Gamut', kind: 'Gamut' }];
+
+  path.forEach((kind, i) => {
+    const prevKind = i ? `${links[i].kind}/` : '';
+
+    const nextLink = {
+      text: kind,
+      kind: `${prevKind}${kind}`,
+    };
+
+    links.push(nextLink);
+  });
+
+  return (
+    <Box
+      display="flex"
+      columnGap="0.5rem"
+      marginBottom="0.5rem"
+      fontWeight="bold"
+      fontSize="16px"
+    >
+      {links.map(({ text, kind }, i) => {
+        const key = `breadcrumb-${text}`;
+        if (i + 1 < links.length) {
+          const kindIndex = storyStore._kinds?.[kind] ? kind : `${kind}/About`;
+
+          return (
+            <React.Fragment key={key}>
+              <SectionLink key={kind} kind={kindIndex}>
+                {text}
+              </SectionLink>
+              <span>&gt;</span>
+            </React.Fragment>
+          );
+        }
+        return <span key={key}>{text}</span>;
+      })}
+    </Box>
+  );
+};
+
+export const Page: React.FC = ({ children }) => {
+  const { kind } = useContext(DocsContext);
+  const {
+    path,
+    status,
+    title,
+    subtitle,
+    parameters: { figmaId, source },
+  } = useKind(kind!);
 
   const npmLink = `https://www.npmjs.com/package/@codecademy/${source}`;
   const figmaLink = `https://www.figma.com/file/${figmaId}`;
@@ -63,13 +105,14 @@ export const Page: React.FC = ({ children }) => {
   return (
     <>
       <Header marginBottom="1rem">
+        <BreadCrumbs path={path} />
         <HeaderRow>
           <HeaderCol>
-            <Title>{pageTitle}</Title>
+            <Title>{title}</Title>
           </HeaderCol>
         </HeaderRow>
         <HeaderRow>
-          {showStatus && (
+          {status && (
             <>
               <HeaderCol>
                 <strong>Status:</strong>
@@ -105,6 +148,7 @@ export const Page: React.FC = ({ children }) => {
           )}
         </HeaderRow>
       </Header>
+      <Description>{subtitle}</Description>
       {children}
     </>
   );