+++ packages/styleguide/__tests__/runSplitStoryshotsTests.ts
@@ -1,8 +1,12 @@
 import initStoryshots, { renderWithOptions } from '@storybook/addon-storyshots';
 import { mount } from 'enzyme';
+import glob from 'glob';
 import path from 'path';
 
-global.STORYSHOTS_TOTAL = 9;
+const STORYSHOT_GLOB = '../__tests__/storyshots/*-test.ts';
+const storyShots = glob.sync(STORYSHOT_GLOB, { cwd: __dirname });
+
+global.STORYSHOTS_TOTAL = storyShots.length;
 
 export const runSplitStoryshotsTests = (index: number) => {
   global.STORYSHOTS_INDEX = index;
@@ -12,7 +16,6 @@ export const runSplitStoryshotsTests = (index: number) => {
     test: renderWithOptions({
       renderer: mount,
     }),
-    storyKindRegex: /^((?!.*?Popover).(?!.*?Overlay).)*$/,
     configPath: path.join(__dirname, '../.storybook'),
   });
 };