+++ packages/eslint-config/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [3.1.5](http://github.com///compare/@codecademy/eslint-config@3.1.4...@codecademy/eslint-config@3.1.5) (2021-01-27)
+
+
+### Reverts
+
+* Revert "A few more disables, from the PRs" ([b7e7855](http://github.com///commit/b7e7855f84c3ca6b8453c92379118aa0d3c45f32))
+
+
+
+### [3.1.4](http://github.com///compare/@codecademy/eslint-config@3.1.3...@codecademy/eslint-config@3.1.4) (2021-01-27)
+
+**Note:** Version bump only for package @codecademy/eslint-config
+
+
+
+
+
+### [3.1.3](http://github.com///compare/@codecademy/eslint-config@3.1.2...@codecademy/eslint-config@3.1.3) (2021-01-22)
+
+**Note:** Version bump only for package @codecademy/eslint-config
+
+
+
+
+
+### [3.1.2](http://github.com///compare/@codecademy/eslint-config@3.1.1...@codecademy/eslint-config@3.1.2) (2021-01-20)
+
+
+### Bug Fixes
+
+* **styleguide:** Enable no jsx-no-undef on mdx ([06b9ceb](http://github.com///commit/06b9cebf86429a825375e8532b1b550a5844ca28))
+
+
+
+### [3.1.1](http://github.com///compare/@codecademy/eslint-config@3.1.0...@codecademy/eslint-config@3.1.1) (2021-01-19)
+
+
+### Bug Fixes
+
+* **MDX:** Parsing Validation and ESlint for MDX WEB-1238 ([a9fc436](http://github.com///commit/a9fc436b415d319bff2858cf75fff6afbb489f6f))
+
+
+
 ## [3.1.0](http://github.com///compare/@codecademy/eslint-config@3.0.7...@codecademy/eslint-config@3.1.0) (2020-12-11)