+++ packages/gamut-icons/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.14.0](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-icons@3.13.0...@codecademy/gamut-icons@3.14.0) (2021-01-27)
+
+
+### Features
+
+* **Icons:** Adds initial Mini Bold streamline icons ([e3c521c](https://github.com/Codecademy/client-modules/commit/e3c521c5e2a4f207db079c942560c54c93125fa5))
+
+
+
+## [3.13.0](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-icons@3.12.2...@codecademy/gamut-icons@3.13.0) (2021-01-12)
+
+
+### Features
+
+* **Icons:** Add video player movie icon ([def3a59](https://github.com/Codecademy/client-modules/commit/def3a59d9c253330bc4703f2c46ded80a1079bac))
+
+
+
 ### [3.12.2](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-icons@3.12.1...@codecademy/gamut-icons@3.12.2) (2020-12-18)